refactor: getSuggest tabId if문 -> switch문 변경 (#99)

+++ server/src/main/java/com/codueon/boostUp/domain/suggest/contoller/SuggestController.java
@@ -71,7 +71,7 @@ public class SuggestController {
 
     @GetMapping("lesson/{lesson-id}/suggest/teacher/tab/{tab-id}")
     public ResponseEntity<MultiResponseDto<?>> getTeacherSuggest(@PathVariable("lesson-id") Long lessonId,
-                                                                 @PathVariable("tab-id") Long tabId,
+                                                                 @PathVariable("tab-id") int tabId,
                                                                  Pageable pageable) {
 
         Long memberId = 1L;
